Tests rework
